refactor: 메인페이지 컴포넌트 rcc/rsc 폴더 분리 (#176)
