+++ src/app/_components/server/home-main-carousel-container.tsx
@@ -1,4 +1,4 @@
-import MainCarousel from '@/components/features/home/home-main-carousel';
+import MainCarousel from '@/app/_components/client/home-main-carousel';
 import carouselData from '@/data/carousel-data.json';
 
 const MainCarouselContainer = () => {
