+++ src/app/_components/server/home-navigation-button.tsx
@@ -1,5 +1,4 @@
 import { ChevronLeft, ChevronRight } from 'lucide-react';
-
 import { NavigationButtonProps } from '@/types/home.carousel.type';
 
 const NavigationButton = ({ direction, onClick }: NavigationButtonProps) => {
