+++ src/app/_components/server/home-weather-header.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { WeatherHeaderProps } from '@/types/home.weather.type';
 
 /**
